LayoutTests/accessibility/mac/meter-subrole-expected.txt


This tests the subrole for meter elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS meter.subrole is 'AXSubrole: AXMeter'
PASS successfullyParsed is true

TEST COMPLETE
